requirements.txt
boost==0.1
cmake==0.7.1
boost-py
face_recognition_models
face_recognition
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
aniso8601==1.2.0
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==1.9
Flask==0.12.1
Flask-Ask==0.9.7
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
pyOpenSSL==17.0.0
python-dateutil==2.7.2
PyYAML==3.12
requests==2.18.4
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
opencv-python
face_recognition 
Django
djangorestframework
request
pandas
django-crispy-forms
